Only blog user can delete blog

+++ part4/blog-list/controllers/blogs.js
@@ -52,6 +52,16 @@ blogsRouter.put('/:id', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' });
+    }
+
+    const blog = await Blog.findById(request.params.id);
+    if (decodedToken.id !== blog.user.toString()) {
+      return response.status(401).json({ error: 'unauthorized to delete this blog' });
+    }
+
     await Blog.findByIdAndDelete(request.params.id);
     response.status(204).end();
   } catch (error) {
